+++ tests/unit/processor/list_comparison/test_list_comparison.py
@@ -21,6 +21,9 @@ from logprep.util.getter import (
 )
 from tests.unit.processor.base import BaseProcessorTestCase
 
+NOT_SET = object()
+"""A sentinel object to indicate that a value has not been provided."""
+
 
 class TestListComparison(BaseProcessorTestCase):
     CONFIG = {
@@ -294,12 +297,24 @@ Hans
         processor.setup()
         assert processor.rules[0].compare_sets == {"bad_users.list": {"Franz", "Heinz", "Hans"}}
 
+    @pytest.mark.parametrize(
+        ("json_content", "content_field"),
+        [
+            pytest.param(["Franz", "Heinz", "Hans"], ""),
+            pytest.param(["Franz", "Heinz", "Hans"], None),
+            pytest.param(
+                ["Franz", "Heinz", "Hans"], NOT_SET, id="no_content_field_entry_in_config"
+            ),
+            pytest.param({"content": ["Franz", "Heinz", "Hans"]}, "content"),
+            pytest.param({"_": ["Franz", "Heinz", "Hans"]}, "_"),
+        ],
+    )
     @responses.activate
-    def test_list_comparison_loads_json_list(self):
+    def test_list_comparison_loads_json_list_from_http(self, json_content, content_field):
         responses.add(
             responses.GET,
             "http://localhost:8080/v2/valuestore/test_4",
-            json.dumps(["Franz", "Heinz", "Hans"]),
+            json.dumps(json_content),
             content_type="application/json",
         )
         rule_dict = {
@@ -311,6 +326,10 @@ Hans
             },
             "description": "",
         }
+
+        if content_field is not NOT_SET:
+            rule_dict["list_comparison"] |= {"content_field": content_field}
+
         config = {
             "type": "list_comparison",
             "rules": [],
@@ -324,6 +343,99 @@ Hans
         processor._rule_tree.add_rule(rule)
         processor.setup()
         assert processor.rules[0].compare_sets == {"bad_users.list": {"Franz", "Heinz", "Hans"}}
+
+    @pytest.mark.parametrize(
+        ("json_content", "content_field"),
+        [
+            pytest.param(["Franz", "Heinz", "Hans"], ""),
+            pytest.param(["Franz", "Heinz", "Hans"], None),
+            pytest.param(
+                ["Franz", "Heinz", "Hans"], NOT_SET, id="no_content_field_entry_in_config"
+            ),
+            pytest.param({"content": ["Franz", "Heinz", "Hans"]}, "content"),
+            pytest.param({"_": ["Franz", "Heinz", "Hans"]}, "_"),
+        ],
+    )
+    def test_list_comparison_loads_json_list_from_file(self, json_content, content_field, tmp_path):
+        file_content = json.dumps(json_content)
+        file_name = "file.json"
+        file_root_path = tmp_path
+        file_path = file_root_path / file_name
+
+        with open(file_path, "w") as f:
+            f.write(file_content)
+
+        rule_dict = {
+            "filter": "user",
+            "list_comparison": {
+                "source_fields": ["user"],
+                "target_field": "user_results",
+                "list_file_paths": [file_name],
+            },
+            "description": "",
+        }
+
+        if content_field is not NOT_SET:
+            rule_dict["list_comparison"] |= {"content_field": content_field}
+
+        config = {
+            "type": "list_comparison",
+            "rules": [],
+            "list_search_base_path": str(file_root_path),
+        }
+
+        HttpGetter._shared.clear()
+
+        processor = Factory.create({"custom_lister": config})
+        rule = processor.rule_class.create_from_dict(rule_dict)
+        processor._rule_tree.add_rule(rule)
+        processor.setup()
+        assert processor.rules[0].compare_sets == {"file.json": {"Franz", "Heinz", "Hans"}}
+
+    @pytest.mark.parametrize(
+        ("json_content", "content_field"),
+        [
+            pytest.param({None: ["Franz", "Heinz", "Hans"]}, ""),
+            pytest.param({None: ["Franz", "Heinz", "Hans"]}, None),
+            pytest.param({"": ["Franz", "Heinz", "Hans"]}, ""),
+            pytest.param({"": ["Franz", "Heinz", "Hans"]}, None),
+        ],
+    )
+    def test_list_comparison_fail_on_json_list_load_from_file(
+        self, json_content, content_field, tmp_path
+    ):
+        file_content = json.dumps(json_content)
+        file_name = "file.json"
+        file_root_path = tmp_path
+        file_path = file_root_path / file_name
+
+        with open(file_path, "w") as f:
+            f.write(file_content)
+
+        rule_dict = {
+            "filter": "user",
+            "list_comparison": {
+                "source_fields": ["user"],
+                "target_field": "user_results",
+                "list_file_paths": [file_name],
+                "content_field": content_field,
+            },
+            "description": "",
+        }
+        config = {
+            "type": "list_comparison",
+            "rules": [],
+            "list_search_base_path": str(file_root_path),
+        }
+
+        HttpGetter._shared.clear()
+
+        processor = Factory.create({"custom_lister": config})
+        rule = processor.rule_class.create_from_dict(rule_dict)
+        processor._rule_tree.add_rule(rule)
+
+        with pytest.raises(ValueError, match="Content is not a list"):
+            processor.setup()
 
     @responses.activate
     def test_list_comparison_loads_rule_using_http_and_updates_with_callback(self, tmp_path):